Create The Coleman-Liau index .py
The Coleman-Liau index was designed to be easily calculated mechanically from samples of hard-copy text. Unlike syllable-based readability indices, it does not require that the character content of words be analyzed, only their length in characters. Therefore, it could be used in conjunction with theoretically simple mechanical scanners that would only need to recognize character, word, and sentence boundaries, removing the need for full optical character recognition or manual keypunching.A score of 6 is 6th grade in the US schooling system. If you are writing for the public, you should aim for a grade level of around 8-10."""